fix: comprehensive codebase improvement — correctness, architecture, hardening
Phase 1 (Correctness): 8 fixes for broken production paths
- DB no longer wiped on every start (remove_file removed, corruption recovery added)
- xray builder credentials extracted from spec_blob via shared parse_settings()
- Speed results no longer truncate at 2.1 Gbps (i32->i64 across models)
- format_now() uses Hinnant civil calendar (was off by days)
- xray Hysteria2 outbound matches upstream ({version:2,address,port})
- Trojan flow field removed (xray-core rejects it), Shadowsocks uot removed
- IP geolocation URL changed http->https, file log writer uses serde_json escape

Phase 2 (Architecture): 4 structural improvements
- to_settings() implemented for all 16 typed ProtocolConfig variants
- gRPC clients unified (XrayGrpcClient+SingBoxGrpcClient -> GrpcStatsClient)
- Dead tcp_ping_concurrency config removed
- speed_test CLIENT_CACHE uses tokio::sync::Mutex (was std sync in async ctx)

Phase 3 (Hardening): 4 safety improvements
- raw .unwrap() replaced with .expect()/let-else in schemex, split_url, common
- Log file path canonicalized (.. traversal rejected)
- updater uses tempfile::tempdir() instead of predictable temp path
- HTTP subscription URLs logged as warning, safe_parse_json with 64KB limit

+++ crates/xray-tui-db/src/models_toasty.rs
@@ -87,7 +87,7 @@ pub struct ProfileExtension {
     pub protocol_id: i64, // was profile_id
 
     pub delay: Option<i32>,
-    pub speed: Option<i32>,
+    pub speed: Option<i64>,
     pub sort_order: Option<i32>,
     pub ip_info: Option<String>,
 
@@ -163,7 +163,7 @@ pub struct PingSession {
     pub ping_type: String,
     pub status: String,
     pub latency_ms: Option<i32>,
-    pub speed_bps: Option<i32>,
+    pub speed_bps: Option<i64>,
     pub ip_info: Option<String>,
     pub error: Option<String>,
     pub created_at: Option<String>,
@@ -180,7 +180,7 @@ pub struct PingResultUpdate {
     pub status: String,
     pub ping_type: String,
     pub latency_ms: Option<i32>,
-    pub speed_bps: Option<i32>,
+    pub speed_bps: Option<i64>,
     pub ip_info: Option<String>,
     pub error: Option<String>,
 }
